Added Profile pic, university logo, fixed profile button placemen
I fixed the placement of the profile button so that it is apart of the main container and not floating around anymore. I also added a placeholder profile pic that displays above the profile button and a uni logo in the bottom right of the container in the hompage and navigation page.